fix(github): support payloads without organization (#76)
adds support for payloads without the repository.organization key from
github. this key isn't actually listed in the official github docs so
i'm not sure how reliablie it is in the long run. if the field isn't
present this allows us to fallback to something that is officially
supported (according to the docs).