chore(release):  Prepare to handle pinned deps. (#169)
* chore(release):  Prepare to handle pinned deps.

Currently, Dinghy is using the vendored packages, but since "private"
Dinghy only has one dependent package -- "Open Core Dinghy", which we
want to trigger changes, we want to be able to set the specific version
of that dependency when building.

This change is a baby step -- it always uses the latest Open Core Dinghy
when compiling, but has the ability (with an env var) to specify the
hash wanted.  When this seems to be working, we'll go back and see how
we might be able to pass in the pinned version.

* Updating vendor dynamically

Made armory-io/monitoring public, so this can now pull down all the
modules needed during build (defaults to master).