fix(templates): templateOrg can be different from target repo (#81)
* fix(templates): templateOrg can be different from target repo

Previously modules were being pulled from the same organization that was
configured for the current repository. That is, if `org-a/project-a`
defined a `dinghyfile` in their repo, they could only source modules
from `org-a`, regardless of what was configured in their `dinghy`
config. This commit fixes this behavior so that the setting is honored.

* chore(startup): allow dinghy port to be overriden by shell var