fix(config): changed two feature flags types from boolean to string. (#184)
When merging two configurations - the default one defined in the application’s code and the user-defined one - if a feature flag is set to false in the user config but set to true in the default config, the default value will take precedence over the user-defined value. In such case, it is impossible to disable the feature.